Initialize new step for dynamic keys
The `StateStepCreator` now initializes new `Step`s after they are
created and only informs the requesting producers about the step's
existence when the initialization is complete. This ensures that only
steps that are ready to process messages are handed off to producers.

The `StateStepCreator` also keeps a list of producers that are waiting
for a particular new step and notifies all of them when the step is
ready. This way we correctly handle a situation where multiple reports
of the same missing key come in before the step to handle that key is
completely set up.